rename content subcommand to manifest-gen, add InterfaceGroup generation, create implementation subcommand and add override flag
