+++ cmd/cli/docs/capact_alpha_manifest-gen.md
@@ -1,0 +1,32 @@
+---
+title: capact alpha manifest-gen
+---
+
+## capact alpha manifest-gen
+
+Manifests generation
+
+### Synopsis
+
+Subcommand for various manifest generation operations
+
+### Options
+
+```
+  -h, --help            help for manifest-gen
+  -o, --output string   Path to the output directory for the generated manifests (default "generated")
+      --override        Override existing manifest files
+```
+
+### Options inherited from parent commands
+
+```
+  -c, --config string   Path to the YAML config file
+```
+
+### SEE ALSO
+
+* [capact alpha](capact_alpha.md)	 - Alpha features
+* [capact alpha manifest-gen implementation](capact_alpha_manifest-gen_implementation.md)	 - Bootstrap new Implementation manifests
+* [capact alpha manifest-gen interface](capact_alpha_manifest-gen_interface.md)	 - Bootstrap new Interface manifests
+
