+++ cmd/cli/docs/capact_alpha_manifest-gen_implementation.md
@@ -1,0 +1,31 @@
+---
+title: capact alpha manifest-gen implementation
+---
+
+## capact alpha manifest-gen implementation
+
+Bootstrap new Implementation manifests
+
+### Synopsis
+
+Bootstrap new Implementation manifests for various tools.
+
+### Options
+
+```
+  -h, --help   help for implementation
+```
+
+### Options inherited from parent commands
+
+```
+  -c, --config string   Path to the YAML config file
+  -o, --output string   Path to the output directory for the generated manifests (default "generated")
+      --override        Override existing manifest files
+```
+
+### SEE ALSO
+
+* [capact alpha manifest-gen](capact_alpha_manifest-gen.md)	 - Manifests generation
+* [capact alpha manifest-gen implementation terraform](capact_alpha_manifest-gen_implementation_terraform.md)	 - Bootstrap Terraform based manifests
+
