+++ internal/cli/alpha/manifestgen/helpers.go
@@ -1,4 +1,4 @@
-package content
+package manifestgen
 
 import (
 	"fmt"
@@ -9,13 +9,22 @@ import (
 	"github.com/pkg/errors"
 )
 
-func writeManifestFiles(files map[string]string) error {
+// WriteManifestFiles writes the manifests file in files parameters to the provided outputDir.
+// Depending on the override parameter is will either override existing manifest files or skip them.
+func WriteManifestFiles(outputDir string, files map[string]string, override bool) error {
+	fmt.Println(override)
+
 	for manifestPath, content := range files {
 		manifestFilepath := strings.ReplaceAll(manifestPath, ".", string(os.PathSeparator)) + ".yaml"
-		outputFilepath := path.Join(manifestOutputDirectory, manifestFilepath)
+		outputFilepath := path.Join(outputDir, manifestFilepath)
 
 		if err := os.MkdirAll(path.Dir(outputFilepath), 0750); err != nil {
 			return errors.Wrap(err, "while creating directory for generated manifests")
+		}
+
+		if _, err := os.Stat(outputFilepath); !override && !os.IsNotExist(err) {
+			fmt.Printf("Skipped %s as it already exists in %s\n", manifestPath, outputFilepath)
+			continue
 		}
 
 		if err := os.WriteFile(outputFilepath, []byte(content), 0600); err != nil {
