+++ internal/cli/alpha/manifestgen/provider.go
@@ -1,4 +1,4 @@
-package content
+package manifestgen
 
 import "fmt"
